requirements.txt
requests
beautifulsoup4
feedgen
python-dotenv
google-generativeai